+++ lib/modules/pool/actions/remove-liquidity/useRemoveLiquidity.tsx
@@ -19,7 +19,7 @@ import { Address } from 'viem'
 import { toHumanAmount } from '../LiquidityActionHelpers'
 import { useDisclosure } from '@chakra-ui/hooks'
 import { TransactionState } from '@/lib/modules/transactions/transaction-steps/lib'
-import { useIterateSteps } from '../useIterateSteps'
+import { useIterateSteps } from '../../../transactions/transaction-steps/useIterateSteps'
 import { useRemoveLiquidityStepConfigs } from './modal/useRemoveLiquidityStepConfigs'
 
 export type UseRemoveLiquidityResponse = ReturnType<typeof _useRemoveLiquidity>
