+++ lib/modules/pool/actions/stake/useStaking.tsx
@@ -12,7 +12,7 @@ import { BPT_DECIMALS } from '../../pool.constants'
 import { useEffect, useState } from 'react'
 import { useTokenApprovalConfigs } from '@/lib/modules/tokens/approvals/useTokenApprovalConfigs'
 import { stakeConfig } from './stakeConfig'
-import { useIterateSteps } from '../useIterateSteps'
+import { useIterateSteps } from '../../../transactions/transaction-steps/useIterateSteps'
 
 export function useStaking() {
   const [humanAmountIn, setHumanAmountIn] = useState<HumanAmountIn | null>(null)
