+++ lib/modules/pool/actions/unstake/useClaimAndUnstakeStepConfigs.tsx
@@ -3,7 +3,7 @@ import { ClaimAndUnstakeButton } from './ClaimAndUnstakeButton'
 import { useHasMinterApproval } from '@/lib/modules/staking/gauge/useHasMinterApproval'
 import { getApproveRelayerConfig } from '@/lib/modules/relayer/approveRelayerConfig'
 import { minterApprovalConfig } from '@/lib/modules/staking/gauge/minterApprovalConfig'
-import { StepConfig } from '../useIterateSteps'
+import { StepConfig } from '../../../transactions/transaction-steps/useIterateSteps'
 import { getChainId } from '@/lib/config/app.config'
 import { GqlChain } from '@/lib/shared/services/api/generated/graphql'
