+++ lib/modules/staking/gauge/minterApprovalConfig.tsx
@@ -1,5 +1,9 @@
 import { TransactionStepButton } from '@/lib/modules/transactions/transaction-steps/TransactionStepButton'
-import { MinterStepProps, OnStepCompleted, StepConfig } from '../../pool/actions/useIterateSteps'
+import {
+  MinterStepProps,
+  OnStepCompleted,
+  StepConfig,
+} from '../../transactions/transaction-steps/useIterateSteps'
 import { useConstructMinterApprovalStep } from './useConstructMinterApprovalStep'
 import { VStack } from '@chakra-ui/react'
 import { GqlChain } from '@/lib/shared/services/api/generated/graphql'
