+++ lib/modules/swap/useSwap.tsx
@@ -20,7 +20,7 @@ import { useSwapStepConfigs } from './useSwapStepConfigs'
 import { TransactionState } from '@/lib/modules/transactions/transaction-steps/lib'
 import { SdkSimulateSwapResponse, SimulateSwapResponse, SwapState } from './swap.types'
 import { SwapHandler } from './handlers/Swap.handler'
-import { useIterateSteps } from '../pool/actions/useIterateSteps'
+import { useIterateSteps } from '../transactions/transaction-steps/useIterateSteps'
 import { isSameAddress } from '@/lib/shared/utils/addresses'
 import { useVault } from '@/lib/shared/hooks/useVault'
 import { NativeWrapUnwrapHandler } from './handlers/NativeWrapUnwrap.handler'
